updated to match UML

+++ src/models/MaintenanceRequest.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class MaintenanceRequest extends Maintenance {
     boolean accepted = false;
 
-    public MaintenanceRequest(Facility facility, String details, int cost, Date date, LocalTime dur, Room room) {
+    public MaintenanceRequest(Facility facility, String details, double cost, Date date, LocalTime dur, Room room) {
         super(facility, details, cost, date, dur, room);
     }
